Restructured Application
[*] Moved application chunks to their respective Blueprints
