+++ app/auth/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('auth', __name__, template_folder='templates')
+
+from app.auth import routes
