+++ app/auth/templates/email/reset_password.txt
