+++ app/auth/templates/login.html
@@ -6,6 +6,6 @@
     <div class="row">
         {{ wtf.quick_form(form) }}
     </div>
-    <p>New User? <a href="{{ url_for('register') }}">Click to Register!</a></p>
-    <p>Forgot Your Password? <a href="{{ url_for('reset_password_request') }}">Click to Reset It</a></p>
+    <p>New User? <a href="{{ url_for('auth.register') }}">Click to Register!</a></p>
+    <p>Forgot Your Password? <a href="{{ url_for('auth.reset_password_request') }}">Click to Reset It</a></p>
 {% endblock %}
