+++ app/auth/templates/registration.html
