+++ app/errors/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('errors', __name__, template_folder='templates')
+
+from app.errors import handlers
