+++ app/errors/handlers.py
@@ -1,14 +1,15 @@
 from flask import render_template
 
-from app import app, db
+from app import db
+from app.errors import bp
 
 
-@app.errorhandler(404)
+@bp.app_errorhandler(404)
 def not_found_error(error):
     return render_template('404.html'), 404
 
 
-@app.errorhandler(500)
+@bp.app_errorhandler(500)
 def internal_error(error):
     db.session.rollback()
     return render_template('500.html'), 500
