+++ app/models.py
@@ -2,11 +2,12 @@ from datetime import datetime
 from hashlib import md5
 from time import time
 
+from flask import current_app
 from flask_login import UserMixin
 import jwt
 from werkzeug.security import generate_password_hash, check_password_hash
 
-from app import app, db, login
+from app import db, login
 
 
 @login.user_loader
@@ -73,13 +74,13 @@ class User(UserMixin, db.Model):
             {
                 'reset_password': self.id,
                 'exp': time() + expires_in
-            }, app.config['SECRET_KEY'], algorithm='HS256'
+            }, current_app.config['SECRET_KEY'], algorithm='HS256'
         ).decode('utf-8')
 
     @staticmethod
     def verify_reset_password_token(token):
         try:
-            id = jwt.decode(token, app.config['SECRET_KEY'], algorithms=['HS256'])['reset_password']
+            id = jwt.decode(token, current_app.config['SECRET_KEY'], algorithms=['HS256'])['reset_password']
         except:
             return
         return User.query.get(id)
