+++ microblog.py
@@ -1,8 +1,9 @@
-from app import app, db
+from app import create_app, db
 from app.models import User, Post
+
+app = create_app()
 
 
 @app.shell_context_processor
 def shell_context():
-
     return {'db': db, 'User': User, 'Post': Post}
